requirements.txt
streamlit==1.29.0
tavily-python==0.1.9
langchain>=0.1.0
langchain-huggingface>=0.0.10
python-dotenv==1.0.1
